Clean up project, add webpack as build tool
First, we cleaned the project (we felt that the .DS_Store files and
.idea folder were unnessecary) and we removed the dist/ folder as it
just contains a copy of everything in the src folder. We added a
postInstall hook with npm that builds a bundle directly from the source
that is exported as the main entrypoint. We used webpack for this
because of the greatly simplified api over gulp. The functionality that
is currently available from gulp tasks should be encapsulated by the npm
scripts `npm run build` and `npm test`.